refactor: extract config resolvers to a separate file
